Add files via upload
Predictive Text Assignment:-

Predictive text allows the user to press a key only once to reach any of the letters it represents. As things stand, the interface, and the standard mode have currently been completely implemented and your task is to complete the implementation of the predictive text mode. Specifically, you must finish off the implementation of the Dictionary component, which is used to store and retrieve words that correspond to sequences of keystrokes.

An example tree is given above in figure 2. This tree encodes five words: “happy”, “hello”, “good”, “god”, and “fun”. Based on this dictionary, pressing the following set of keys: 3,8,6 would result in the word “fun” being selected. In standard mode, this sequence of keystrokes would result in “dtm”. To get “fun” in standard mode, you would have had to do the following: press 3 (3 times), press 8 (2 times), and press 6 (2 times).

Also, in predictive mode, pressing 4 would result in “